fix(rename): Rename NotAuthorized to UnauthorizedAccess
