--- packages/module/src/NotAuthorized/index.ts
@@ -1,2 +1,0 @@
-export { default } from './NotAuthorized'
-export * from './NotAuthorized'
