+++ packages/module/src/UnauthorizedAccess/UnauthorizedAccess.test.tsx
@@ -1,39 +1,39 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import { QuestionCircleIcon } from '@patternfly/react-icons';
-import NotAuthorized from './NotAuthorized';
+import UnauthorizedAccess from './UnauthorizedAccess';
 
-describe('NotAuthorized component', () => {
+describe('UnauthorizedAccess component', () => {
   const initialProps = {
     serviceName: 'Foo',
   };
   it('should render', () => {
-    const { container } = render(<NotAuthorized {...initialProps} />);
+    const { container } = render(<UnauthorizedAccess {...initialProps} />);
     expect(container.firstChild).toMatchSnapshot();
   });
 
   it('should apply custom styles', () => {
-    const { container } = render(<NotAuthorized {...initialProps} className="something" />);
+    const { container } = render(<UnauthorizedAccess {...initialProps} className="something" />);
     expect(container.firstChild).toMatchSnapshot();
   });
 
   it('should use custom icon', () => {
-    const { container } = render(<NotAuthorized {...initialProps} icon={QuestionCircleIcon} />);
+    const { container } = render(<UnauthorizedAccess {...initialProps} icon={QuestionCircleIcon} />);
     expect(container.firstChild).toMatchSnapshot();
   });
 
   it('should not show buttons', () => {
-    const { container } = render(<NotAuthorized {...initialProps} showReturnButton={false} />);
+    const { container } = render(<UnauthorizedAccess {...initialProps} showReturnButton={false} />);
     expect(container.firstChild).toMatchSnapshot();
   });
 
   it('should show custom body text', () => {
-    const { container } = render(<NotAuthorized {...initialProps} bodyText="Some text" />);
+    const { container } = render(<UnauthorizedAccess {...initialProps} bodyText="Some text" />);
     expect(container.firstChild).toMatchSnapshot();
   });
 
   it('should show custom title', () => {
-    const { container } = render(<NotAuthorized titleText="Custom title" />);
+    const { container } = render(<UnauthorizedAccess titleText="Custom title" />);
     expect(container.firstChild).toMatchSnapshot();
   });
 
@@ -50,7 +50,7 @@ describe('NotAuthorized component', () => {
         3
       </button>
     ];
-    const { container } = render(<NotAuthorized {...initialProps} primaryAction={primaryAction} secondaryActions={secondaryActions} />);
+    const { container } = render(<UnauthorizedAccess {...initialProps} primaryAction={primaryAction} secondaryActions={secondaryActions} />);
     expect(container.firstChild).toMatchSnapshot();
   });
 });
