+++ packages/module/src/UnauthorizedAccess/index.ts
@@ -1,0 +1,2 @@
+export { default } from './UnauthorizedAccess';
+export * from './UnauthorizedAccess';
